+++ webserver/Websockets.cpp
@@ -69,7 +69,7 @@ namespace http {
 					masking_key[i] = rand();
 					res += masking_key[i];
 				}
-				res += unmask(masking_key, (const uint8_t *)payload.c_str(), payloadlen);
+				res += unmask(masking_key, (const uint8_t *)payload.c_str(), (size_t)payloadlen);
 			}
 			else {
 				res += payload;
